src/requirements.txt
beautifulsoup4
requests
facemorpher
pyyaml
